backend/requirements.txt
alembic==1.12.1
annotated-types==0.7.0
anthropic==0.7.0
anyio==3.7.1
certifi==2025.11.12
click==8.3.1
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.104.1
filelock==3.20.1
fsspec==2025.12.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.2.3
idna==3.11
jieba==0.42.1
jiter==0.12.0
joblib==1.5.3
Mako==1.3.10
MarkupSafe==3.0.3
numpy==1.26.2
openai==2.14.0
packaging==25.0
pandas==2.1.3
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
pypinyin==0.49.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.3
scikit-learn==1.3.2
scipy==1.16.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
threadpoolctl==3.6.0
tokenizers==0.22.1
tqdm==4.67.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uvicorn==0.24.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
